ДОБРА КУРСОВА/requirements.txt
# ============================================
# ОБОВ'ЯЗКОВІ ЗАЛЕЖНОСТІ (Core Dependencies)
# ============================================

# Базові бібліотеки для роботи з даними
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Препроцесинг тексту
nltk>=3.6.0

# Візуалізація даних
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Управління змінними оточення
python-dotenv>=0.19.0

# Gradio інтерфейс для демо
gradio>=3.40.0

# Додаткові утиліти
tqdm>=4.62.0

# ============================================
# ОПЦІОНАЛЬНІ ЗАЛЕЖНОСТІ (для роботи з ML моделями)
# Встановлюйте тільки якщо плануєте працювати з notebooks 03-06
# ============================================

# Для генерації ембедингів (notebooks 03-06)
# sentence-transformers>=2.2.0
# transformers>=4.20.0

# Для зменшення розмірності (notebook 03)
# umap-learn>=0.5.3

# ============================================
# JUPYTER NOTEBOOKS
# ============================================

# Для роботи з Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# ============================================
# ІНСТРУКЦІЇ ДЛЯ ВСТАНОВЛЕННЯ
# ============================================

# Базова установка (швидко):
#   pip install -r requirements.txt
#
# Для роботи з ML моделями (додатково):
#   pip install sentence-transformers transformers umap-learn
#
# CPU версія PyTorch (якщо torch встановлюється занадто довго):
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
#   pip install sentence-transformers transformers umap-learn
